# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "my project doc",
        "description": "\u6e38\u620f\u6570\u636e\u8bc4\u4f30\u63a5\u53e3\u6587\u6863, V1.0<br>",
        "version": "1.0"
    },
    "host": "assess.act.sdo.com:81",
    "schemes": [
        "http"
    ],
    "consumes": [
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/index.php/TurnoverEvaluateModel/currentEvaluate": {
            "get": {
                "tags": [
                    "TurnoverEvaluationModel"
                ],
                "summary": "\u9884\u4f30\u6d41\u6c34",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u6e38\u620f\u540d\u79f0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u6e38\u620f\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ip",
                        "in": "query",
                        "description": "\u662f\u5426\u6709ip",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "market_cost",
                        "in": "query",
                        "description": "\u5e02\u573a\u8d39\u91d1\u989d",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "estimated_quantity",
                        "in": "query",
                        "description": "\u9884\u5b9a\u91cf",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "1dayret",
                        "in": "query",
                        "description": "\u6b21\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayret",
                        "in": "query",
                        "description": "7\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "30dayret",
                        "in": "query",
                        "description": "30\u65e5\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayltv",
                        "in": "query",
                        "description": "7\u5929LTV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "arpu",
                        "in": "query",
                        "description": "ARPV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "modified_scale",
                        "in": "query",
                        "description": "\u4fee\u6539\u540e\u7684\u7528\u6237\u89c4\u6a21",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f",
                        "schema": {
                            "properties": {
                                "first_month_turnover": {
                                    "description": "\u9996\u6708\u6d41\u6c34",
                                    "type": "integer"
                                },
                                "third_month_turnover": {
                                    "description": "\u4e09\u6708\u6d41\u6c34",
                                    "type": "integer"
                                },
                                "six_month_turnover": {
                                    "description": "\u534a\u5e74\u6d41\u6c34",
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/index.php/TurnoverEvaluateModel/profitRateEvaluate": {
            "get": {
                "tags": [
                    "TurnoverEvaluationModel"
                ],
                "summary": "\u9884\u4f30\u5229\u6da6\u7387",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u6e38\u620f\u540d\u79f0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u6e38\u620f\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ip",
                        "in": "query",
                        "description": "\u662f\u5426\u6709ip",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "turnover",
                        "in": "query",
                        "description": "\u6d41\u6c34",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "copyright_fee",
                        "in": "query",
                        "description": "\u7248\u6743\u91d1",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "agent_share",
                        "in": "query",
                        "description": "\u4ee3\u7406\u5206\u6210",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "dev_share",
                        "in": "query",
                        "description": "\u7814\u53d1\u5206\u6210",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "ios_income_ratio",
                        "in": "query",
                        "description": "ios\u6536\u5165\u6bd4\u4f8b",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "android_income_ratio",
                        "in": "query",
                        "description": "\u5b89\u5353\u6e20\u9053\u6536\u5165\u6bd4\u4f8b",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "gplus_income_ratio",
                        "in": "query",
                        "description": "\u5b98\u65b9G+\u6536\u5165\u6bd4\u4f8b",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "idc_cost",
                        "in": "query",
                        "description": "IDC\u6210\u672c",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "headcount",
                        "in": "query",
                        "description": "\u4eba\u5458\u6570\u91cf",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f",
                        "schema": {
                            "properties": {
                                "first_month_profit_ration": {
                                    "description": "\u4e00\u4e2a\u6708\u5229\u6da6\u7387",
                                    "type": "integer"
                                },
                                "six_month_profit_ration": {
                                    "description": "\u534a\u5e74\u5229\u6da6\u7387",
                                    "type": "integer"
                                },
                                "first_year_profit_ration": {
                                    "description": "\u4e00\u5e74\u5229\u6da6\u7387",
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/index.php/TurnoverEvaluateModel/marketFeeEvaluate": {
            "get": {
                "tags": [
                    "TurnoverEvaluationModel"
                ],
                "summary": "\u9884\u4f30\u6700\u6709\u5e02\u573a\u8d39\u91d1\u989d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u6e38\u620f\u540d\u79f0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u6e38\u620f\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ip",
                        "in": "query",
                        "description": "\u662f\u5426\u6709ip",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "estimated_quantity",
                        "in": "query",
                        "description": "\u9884\u5b9a\u91cf",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "1dayret",
                        "in": "query",
                        "description": "\u6b21\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayret",
                        "in": "query",
                        "description": "7\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "30dayret",
                        "in": "query",
                        "description": "30\u65e5\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayltv",
                        "in": "query",
                        "description": "7\u5929LTV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "arpu",
                        "in": "query",
                        "description": "ARPV",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f",
                        "schema": {
                            "properties": {
                                "efficiency_market_fee": {
                                    "description": "\u6700\u4f18\u5e02\u573a\u8d39",
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}